separate config data and status data and extract common
