+++ Project/common/publicDefine.h
@@ -1,0 +1,16 @@
+#ifndef _PUBLICDEFINE_H
+#define _PUBLICDEFINE_H
+
+#include <stm8s.h> //Required for the stdint typedefs
+#include "stdio.h"
+#include "string.h"
+#include "stm8s_conf.h"
+
+// Common Data Type
+#define UC                        uint8_t
+#define US                        uint16_t
+#define UL                        uint32_t
+#define SHORT                     int16_t
+#define LONG                      int32_t
+
+#endif /* _PUBLICDEFINE_H */
