+++ Project/src/ProtocolParser.c
@@ -105,7 +105,6 @@ uint8_t ParseProtocol(){
         { // request nodeid response
           gConfig.nodeID = lv_nodeID;
           memcpy(gConfig.NetworkID, rcvMsg.payload.data, sizeof(gConfig.NetworkID));
-          gIsChanged = TRUE;
           GotNodeID();
           // Increase brightness to indicate ID required
           if( gConfig.cntRFReset < MAX_RF_RESET_TIME ) {
@@ -173,7 +172,7 @@ uint8_t ParseProtocol(){
         gConfig.pirLevel[1] = rcvMsg.payload.data[1];
         break;
       }
-      gIsChanged = TRUE;
+      gIsConfigChanged = TRUE;
       Msg_NodeConfigAck(_sender, _sensor);
       return 1;
     }else if( _type == I_GET_NONCE ) {
@@ -706,7 +705,7 @@ void MsgScanner_ConfigAck(uint8_t offset,uint8_t cfglen,bool _isByUniqueid) {
 void Process_SetConfig(u8 _len) {
   uint8_t offset = rcvMsg.payload.data[1];
   memcpy((void *)((uint16_t)(&gConfig) + offset),rcvMsg.payload.data+2,_len);
-  gIsChanged = TRUE;
+  gIsConfigChanged = TRUE;
 }
 //////set config by uniqueid data struct/////////////////////
 //typedef struct
@@ -721,7 +720,7 @@ void Process_SetConfig(u8 _len) {
 void Process_SetDevConfig(u8 _len) {
     uint8_t offset = rcvMsg.payload.data[1];
     memcpy((void *)((uint16_t)(&gConfig) + offset),rcvMsg.payload.data+2+UNIQUE_ID_LEN,_len);
-    gIsChanged = TRUE;
+    gIsConfigChanged = TRUE;
 }
 bool IsNodeidValid(uint8_t nodeid)
 {
@@ -825,7 +824,7 @@ void Process_SetupRF(const UC *rfData,uint8_t rflen)
     gResetNode = TRUE;
   if(gResetNode || gResetRF || bNeedChangeCfg)
   {
-    gIsChanged = TRUE;
+    gIsConfigChanged = TRUE;
   }
 }
 //----------------------------------------------
